Оптимизация классов
