+++ src/test/java/UserUpdateTest.java
@@ -9,7 +9,6 @@ import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import request.User;
 import response.Message;
-import response.UserMessage;
 
 import static org.apache.http.HttpStatus.SC_OK;
 import static org.apache.http.HttpStatus.SC_UNAUTHORIZED;
@@ -66,7 +65,7 @@ public class UserUpdateTest {
     public void updateUserWithAuthorization() {
         ValidatableResponse response = client.updateInfo(user);
         int statusCode = response.extract().statusCode();
-        UserMessage message = response.extract().as(UserMessage.class);
+        Message message = response.extract().as(Message.class);
 
         assertThat(statusCode, equalTo(SC_OK));
         message.check(user.getName(), user.getEmail(), true);
